requirements.txt
asgiref==3.5.2
autopep8==1.6.0
backports.zoneinfo==0.2.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
defusedxml==0.7.1
dj-database-url==0.5.0
Django==4.0.6
django-phonenumber-field==6.3.0
django-phonenumbers==1.0.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
djoser==2.1.0
drf-yasg==1.21.3
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
oauthlib==3.2.0
packaging==21.3
phonenumbers==8.12.52
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.10
whitenoise==6.2.0
